refactor(tools,package-manager): split fix-version selection and lockfile-sync into their own modules (4.4.5 holistic-fixes)
osv.ts and package-manager.ts crossed the 500-line budget with this
round's additions and the self-guardrail flagged both as net-new large
files. The pure fix-version selection (fixed-event extraction, the
per-installation resolution key) moves to osv-fix-select.ts and the
lockfile-sync check plus resync install move to
package-manager-lockfile.ts; osv.ts re-exports so consumers keep one
import, and the lockfile consumers import the new module directly.
No behavior change.
